+++ src/main/java/com/example/Price.java
@@ -1,12 +1,12 @@
 package com.example;
 
 public class Price {
-    public int value;
+    public Double value;
     public String currency_code;
 
     public Price() {}
 
-    public Price(int value, String currency_code) {
+    public Price(Double value, String currency_code) {
         this.value = value;
         this.currency_code = currency_code;
     }
